requirements.txt
beautifulsoup4==4.11.1
chromadb==0.5.3
duckduckgo_search==6.1.0
fastapi==0.111.0
fitz==0.0.1.dev2
Flask==2.3.1
googlesearch_python==1.2.4
langchain==0.2.
langchain_community==0.2.6
langchain_core==0.2.10
langchain_google_genai==1.0.7
langdetect==1.0.9
numpy==2.0.0
pandas==1.5.2
pdfplumber==0.11.0
Pillow==10.3.0
pinecone==4.0.0
plotly==5.22.0
protobuf==5.27.2
pydantic==1.10.9
PyPDF2==3.0.1
Requests==2.32.3
streamlit==1.35.0
torch==2.2.0
tqdm==4.66.4
transformers==4.41.2
